go modules, config file support
Code is split up in multiple files for ease of use and readability instead of everything being in 1 huge main.go file.

Support for .config files
