+++ .gitignore
@@ -1,1 +1,1 @@
-main
+fileshare
