+++ go.mod
@@ -1,0 +1,5 @@
+module fileshare
+
+go 1.23.4
+
+require gopkg.in/yaml.v3 v3.0.1 // indirect
